pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=2.0.2",
    "pylint>=2.17.2",
    "pytest>=7.3.1",
]

[project]
name = "micro-videos"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo Figueiredo", email = "gustavohmfigueiredo@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
test="pytest src"
test_cov="pytest src --cov ./src"
test_html="pytest src --cov ./src --cov-report html:./__coverage"
ci_test="pytest src --cov ./src --cov-fail-under 80"
lint="pylint ./src"
formatter="autopep8 --in-place --recursive ./src"

[tool.autopep8]
max_line_length = 100
in-place = true
recursive = true

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
